refactor: move templates to it's own package
