+++ internal/controllers/index.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 
 	"github.com/gofiber/fiber/v2"
-	"github.com/olukkas/ushort/internal/controllers/templates"
+	"github.com/olukkas/ushort/internal/templates"
 )
 
 func HelloRoute(c *fiber.Ctx) error {
